fix: move encrypted icon to excalidraw-app
