+++ src/excalidraw-app/index.scss
@@ -1,0 +1,14 @@
+.excalidraw {
+  .encrypted-icon {
+    position: absolute;
+    bottom: 13px;
+    left: 12.5rem;
+    border-radius: var(--space-factor);
+    color: var(--icon-green-fill-color);
+
+    svg {
+      width: 1.2rem;
+      height: 1.2rem;
+    }
+  }
+}
