+++ src/excalidraw-app/index.tsx
@@ -52,6 +52,11 @@ import {
 } from "./data/localStorage";
 import CustomStats from "./CustomStats";
 import { RestoredDataState } from "../data/restore";
+import clsx from "clsx";
+import { Tooltip } from "../components/Tooltip";
+import { shield } from "../components/icons";
+
+import "./index.scss";
 
 const languageDetector = new LanguageDetector();
 languageDetector.init({
@@ -324,7 +329,23 @@ const ExcalidrawWrapper = () => {
   );
 
   const renderFooter = useCallback(
-    (isMobile: boolean) => {
+    (isMobile: boolean, appState: AppState) => {
+      const renderEncryptedIcon = () => (
+        <a
+          className={clsx("encrypted-icon tooltip zen-mode-visibility", {
+            "zen-mode-visibility--hidden": appState.zenModeEnabled,
+          })}
+          href="https://blog.excalidraw.com/end-to-end-encryption/"
+          target="_blank"
+          rel="noopener noreferrer"
+          aria-label={t("encrypted.link")}
+        >
+          <Tooltip label={t("encrypted.tooltip")} position="above" long={true}>
+            {shield}
+          </Tooltip>
+        </a>
+      );
+
       const renderLanguageList = () => (
         <LanguageList
           onChange={(langCode) => {
@@ -367,7 +388,12 @@ const ExcalidrawWrapper = () => {
           </div>
         );
       }
-      return renderLanguageList();
+      return (
+        <>
+          {renderEncryptedIcon()}
+          {renderLanguageList()}
+        </>
+      );
     },
     [langCode],
   );
